- since the write buffer is tracked in stdf_file, dont bother creating a local one in each write function - simplify header init in each write function - make sure we dont overflow the write buffer - declare local functions static as they should be

+++ src/libstdf.c
@@ -434,7 +434,7 @@ int stdf_close(stdf_file *file)
 {
 	int ret, ret_errno;
 	if (file->__output) {
-		_stdf_write_flush(file, -1);
+		_stdf_write_flush(file, (size_t)-1);
 		free(file->__output);
 	}
 	ret = file->fops->close(file);
